chore(declarations): move declarations from types/ to declarations.d.ts
