--- types/graceful-goodbye/index.d.ts
@@ -1,3 +1,0 @@
-declare module 'graceful-goodbye' {
-  export = function (cb: (...args: any[]) => {}) {};
-}
