--- types/random-access/index.d.ts
@@ -1,9 +1,0 @@
-declare module 'random-access-memory' {
-  function foo(): any;
-  export = foo;
-}
-
-declare module 'random-access-web' {
-  function foo(): any;
-  export = foo;
-}
