--- types/randombytes/index.d.ts
@@ -1,3 +1,0 @@
-declare module 'randombytes' {
-  export = randombytes;
-}
