--- types/safet-catch/index.d.ts
@@ -1,4 +1,0 @@
-declare module 'safety-catch' {
-  function safety(error: any): any;
-  export = safety;
-}
