--- types/sodium-universal/index.d.ts
@@ -1,4 +1,0 @@
-declare module 'sodium-universal' {
-  import * as sodium from 'sodium-native';
-  export = sodium;
-}
