--- types/turbo-hash-map/index.d.ts
@@ -1,3 +1,0 @@
-declare module 'turbo-hash-map' {
-  export = Map;
-}
